# cmd/service/gateway/docs/swagger.yaml
basePath: /api/v1
definitions:
  models.AuthenticationRes:
    properties:
      email:
        type: string
      jwt:
        type: string
    type: object
  models.CreateUserReq:
    properties:
      email:
        type: string
      name:
        type: string
      password:
        type: string
    type: object
  models.ErrorResponse:
    properties:
      code:
        type: string
      message:
        type: string
    type: object
  models.LoginReq:
    properties:
      email:
        type: string
      password:
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: nguyenanhthong1995@gmail.com
    name: API Support
    url: http://www.swagger.io/support
  description: This is a server build by golang.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Thong API
  version: "1.0"
paths:
  /auth/login:
    post:
      consumes:
      - application/json
      description: Login user
      parameters:
      - description: Create order
        in: body
        name: CreateUserReq
        required: true
        schema:
          $ref: '#/definitions/models.LoginReq'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.AuthenticationRes'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Login user
  /auth/register:
    post:
      consumes:
      - application/json
      description: register user
      parameters:
      - description: Create order
        in: body
        name: CreateUserReq
        required: true
        schema:
          $ref: '#/definitions/models.CreateUserReq'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.CreateUserReq'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Register user
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
  BasicAuth:
    type: basic
swagger: "2.0"